implement benchmark mechanics
